Securing Admin Feature Views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -9,7 +10,6 @@ from .forms import ProductForm
 
 def all_products(request):
     """ A view to return the all products, sorting & searching"""
-
     products = Product.objects.all()
     query = None
     categories = None
@@ -70,8 +70,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Admin add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this feature is for Admin users only')
+        return redirect(reverse('products'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,8 +98,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this feature is for Admin users only')
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -118,8 +128,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this feature is for Admin users only')
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
